Move change_dropout_strength to utils in evaluation
